% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteriobrasil-package.R
\docType{package}
\name{criteriobrasil-package}
\alias{criteriobrasil}
\alias{criteriobrasil-package}
\title{criteriobrasil}
\description{
Tidy tables from the Brazilian Economic Classification Criterion (CCEB),
including scoring rules, class cutoffs, income estimates, and geographic
distributions.
}
\author{
\strong{Maintainer}: Vinicius Oike \email{viniciusoike@gmail.com}

}
\keyword{internal}
